refactor: move LocalSendClient to src/core/send.ts
